+++ docs/conf.py
@@ -169,6 +169,7 @@ html_permalinks_icon = Icons.permalinks_icon
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+html_js_files = ["gurubase-widget.js"]  # gurubase AI widget
 
 # Add any extra paths that contain custom files (such as robots.txt or
 # .htaccess) here, relative to this directory. These files are copied